dashboard/backend/requirements.txt
# Flask Backend Dependencies
flask==3.0.0
flask-cors==4.0.0

# MQTT Client
paho-mqtt==1.6.1

# OPC UA Client
opcua==0.98.13

# Firebase
firebase-admin==6.2.0

# Machine Learning
scikit-learn==1.3.0
joblib==1.3.2
numpy==1.24.3
pandas==2.0.3

# Utilities
python-dotenv==1.0.0